#317 Moving templates
