+++ src/GameCards/MlbTheShowForecaster.GameCards.Infrastructure/Services/Reports/Templates/ForecastImpact.template.html
