+++ src/GameCards/MlbTheShowForecaster.GameCards.Infrastructure/Services/Reports/Templates/Layout.template.html
