+++ src/GameCards/MlbTheShowForecaster.GameCards.Infrastructure/Services/Reports/Templates/PlayerCardForecast.template.html
