+++ tests/GameCards/MlbTheShowForecaster.GameCards.Infrastructure.Tests/Services/Reports/TestFiles/StringReplacementForecastGenerator_Expected.html
